+++ Tests/XcodeProjTests/Project/WorkspaceSettingsTests.swift
@@ -23,6 +23,20 @@ final class WorkspaceSettingsTests: XCTestCase {
         XCTAssertTrue(got.autoCreateSchemes == true)
     }
 
+    func test_init_when_relative_derivedData_is_enabled() throws {
+        let path = fixturesPath() + "WorkspaceSettings/OriginalRelativeDerivedData.xcsettings"
+        let got = try WorkspaceSettings.at(path: path)
+        XCTAssertTrue(got.derivedDataCustomLocation == "CustomizedDerivedData")
+        XCTAssertTrue(got.derivedDataLocationStyle == .workspaceRelativePath)
+    }
+    
+    func test_init_when_absolute_derivedData_is_enabled() throws {
+        let path = fixturesPath() + "WorkspaceSettings/OriginalAbsoluteDerivedData.xcsettings"
+        let got = try WorkspaceSettings.at(path: path)
+        XCTAssertTrue(got.derivedDataCustomLocation == "/User/xcodeproj/DerivedData")
+        XCTAssertTrue(got.derivedDataLocationStyle == .absolutePath)
+    }
+
     func test_equals() {
         let lhs = WorkspaceSettings(buildSystem: .new)
         let rhs = WorkspaceSettings(buildSystem: .original)
@@ -36,10 +50,14 @@ final class WorkspaceSettingsTests: XCTestCase {
 
             var settings = try WorkspaceSettings.at(path: path)
             settings.buildSystem = .original
+            settings.derivedDataLocationStyle = .workspaceRelativePath
+            settings.derivedDataCustomLocation = "DerivedData"
             try settings.write(path: copyPath, override: true)
 
             settings = try WorkspaceSettings.at(path: copyPath)
             XCTAssertEqual(settings.buildSystem, .original)
+            XCTAssertEqual(settings.derivedDataLocationStyle, .workspaceRelativePath)
+            XCTAssertEqual(settings.derivedDataCustomLocation, "DerivedData")
         }
     }
 }